redesign PCB to fit (more) Micro ATX cases
The original PCB hung off the side of the motherboard.
If the motherboard is mounted very close to the power
supply, the PCB won't physically fit.

One example of a case with this issue is the Rosewill
LINE-M.

The redesigned PCB should (hopefully) not hang off
edge of the motherboard at all.  I don't believe it
will physically obstruct any other connector, but I
haven't had it fabbed yet, so we'll see.